SIIC/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5.5
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
distlib==0.3.0
Django==3.1.7
django-braces==1.14.0
django-cleanup==5.1.0
django-crispy-forms==1.11.1
django-fontawesome==1.0
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
idna==2.9
iniconfig==1.0.1
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.3
jupyter-client==6.1.11
jupyter-core==4.7.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.812
mypy-extensions==0.4.3
packaging==20.3
parso==0.6.2
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pylint==2.4.4
pyodbc==4.0.30
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.23.0
six==1.14.0
sqlparse==0.4.1
toml==0.10.1
tornado==6.1
traitlets==4.3.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.8
virtualenv==20.0.10
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.11.2
